requirements.txt
asgiref==3.4.1
Django==3.2.18
django-tinymce==3.6.1
django-versatileimagefield==3.0
Pillow==8.4.0
pkg_resources==0.0.0
python-decouple==3.8
python-magic==0.4.27
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1
